Moved CTags preprocess subroutine in proper place
