+++ legacy/builder/preprocess_sketch.go
@@ -88,5 +88,5 @@ func PreprocessSketchWithArduinoPreprocessor(ctx *types.Context) error {
 
 	//fmt.Printf("PREPROCESSOR OUTPUT:\n%s\n", output)
 	ctx.SketchSourceAfterArduinoPreprocessing = string(result)
-	return bldr.SketchSaveItemCpp(ctx.Sketch.MainFile, []byte(ctx.SketchSourceAfterArduinoPreprocessing), ctx.SketchBuildPath)
+	return bldr.SketchSaveItemCpp(ctx.Sketch.MainFile, result, ctx.SketchBuildPath)
 }
